move persistence to sqlite module
 - move authority into store::sqlite
 - rename Authority to SqliteAuthority
 - make SqliteAuthority implement Authority
