+++ tests/integration-tests/tests/catalog_tests.rs
@@ -11,13 +11,14 @@ use trust_dns::rr::*;
 use trust_dns::serialize::binary::{BinDecodable, BinEncodable};
 
 use trust_dns_server::authority::*;
+use trust_dns_server::store::sqlite::SqliteAuthority;
 
 use trust_dns_integration::authority::create_example;
 use trust_dns_integration::*;
 
-pub fn create_test() -> Authority {
+pub fn create_test() -> SqliteAuthority {
     let origin: Name = Name::parse("test.com.", None).unwrap();
-    let mut records: Authority = Authority::new(
+    let mut records: SqliteAuthority = SqliteAuthority::new(
         origin.clone(),
         BTreeMap::new(),
         ZoneType::Master,
@@ -120,8 +121,8 @@ fn test_catalog_lookup() {
     let test_origin = test.origin().clone();
 
     let mut catalog: Catalog = Catalog::new();
-    catalog.upsert(origin.clone(), example);
-    catalog.upsert(test_origin.clone(), test);
+    catalog.upsert(origin.clone(), Box::new(example));
+    catalog.upsert(test_origin.clone(), Box::new(test));
 
     let mut question: Message = Message::new();
 
@@ -202,7 +203,7 @@ fn test_catalog_nx_soa() {
     let origin = example.origin().clone();
 
     let mut catalog: Catalog = Catalog::new();
-    catalog.upsert(origin.clone(), example);
+    catalog.upsert(origin.clone(), Box::new(example));
 
     let mut question: Message = Message::new();
 
@@ -264,7 +265,7 @@ fn test_axfr() {
         ))).clone();
 
     let mut catalog: Catalog = Catalog::new();
-    catalog.upsert(origin.clone(), test);
+    catalog.upsert(origin.clone(), Box::new(test));
 
     let mut query: Query = Query::new();
     query.set_name(origin.clone().into());
@@ -379,7 +380,7 @@ fn test_axfr_refused() {
     let origin = test.origin().clone();
 
     let mut catalog: Catalog = Catalog::new();
-    catalog.upsert(origin.clone(), test);
+    catalog.upsert(origin.clone(), Box::new(test));
 
     let mut query: Query = Query::new();
     query.set_name(origin.clone().into());
