+++ tests/integration-tests/tests/client_future_tests.rs
@@ -38,7 +38,7 @@ use trust_dns_proto::error::ProtoError;
 use trust_dns_proto::xfer::DnsResponse;
 #[cfg(feature = "dnssec")]
 use trust_dns_proto::xfer::{DnsMultiplexer, DnsMultiplexerConnect, DnsMultiplexerSerialResponse};
-use trust_dns_server::authority::Catalog;
+use trust_dns_server::authority::{Authority, Catalog};
 
 use trust_dns_integration::authority::create_example;
 use trust_dns_integration::{NeverReturnsClientStream, TestClientStream};
@@ -47,7 +47,7 @@ use trust_dns_integration::{NeverReturnsClientStream, TestClientStream};
 fn test_query_nonet() {
     let authority = create_example();
     let mut catalog = Catalog::new();
-    catalog.upsert(authority.origin().clone(), authority);
+    catalog.upsert(authority.origin().clone(), Box::new(authority));
 
     let mut io_loop = Runtime::new().unwrap();
     let (stream, sender) = TestClientStream::new(Arc::new(Mutex::new(catalog)));
@@ -189,7 +189,7 @@ where
 fn test_notify() {
     let authority = create_example();
     let mut catalog = Catalog::new();
-    catalog.upsert(authority.origin().clone(), authority);
+    catalog.upsert(authority.origin().clone(), Box::new(authority));
 
     let mut io_loop = Runtime::new().unwrap();
     let (stream, sender) = TestClientStream::new(Arc::new(Mutex::new(catalog)));
@@ -260,7 +260,7 @@ fn create_sig0_ready_client(
 
     // setup the catalog
     let mut catalog = Catalog::new();
-    catalog.upsert(authority.origin().clone(), authority);
+    catalog.upsert(authority.origin().clone(), Box::new(authority));
 
     let signer = Arc::new(signer);
     let (stream, sender) = TestClientStream::new(Arc::new(Mutex::new(catalog)));
