+++ tests/integration-tests/tests/lookup_tests.rs
@@ -21,6 +21,7 @@ use trust_dns_resolver::lookup_ip::LookupIpFuture;
 use trust_dns_resolver::lookup_state::CachingClient;
 use trust_dns_resolver::Hosts;
 use trust_dns_server::authority::{Authority, Catalog};
+use trust_dns_server::store::sqlite::SqliteAuthority;
 
 use trust_dns_integration::authority::create_example;
 use trust_dns_integration::mock_client::*;
@@ -30,7 +31,7 @@ use trust_dns_integration::TestClientStream;
 fn test_lookup() {
     let authority = create_example();
     let mut catalog = Catalog::new();
-    catalog.upsert(authority.origin().clone(), authority);
+    catalog.upsert(authority.origin().clone(), Box::new(authority));
 
     let mut io_loop = Runtime::new().unwrap();
     let (stream, sender) = TestClientStream::new(Arc::new(Mutex::new(catalog)));
@@ -61,7 +62,7 @@ fn test_lookup() {
 fn test_lookup_hosts() {
     let authority = create_example();
     let mut catalog = Catalog::new();
-    catalog.upsert(authority.origin().clone(), authority);
+    catalog.upsert(authority.origin().clone(), Box::new(authority));
 
     let mut io_loop = Runtime::new().unwrap();
     let (stream, sender) = TestClientStream::new(Arc::new(Mutex::new(catalog)));
@@ -81,7 +82,7 @@ fn test_lookup_hosts() {
         RecordType::A,
         Lookup::new_with_max_ttl(
             Query::query(Name::from_str("www.example.com.").unwrap(), RecordType::A),
-            Arc::new(vec![RData::A(Ipv4Addr::new(10, 0, 1, 104))])
+            Arc::new(vec![RData::A(Ipv4Addr::new(10, 0, 1, 104))]),
         ),
     );
 
@@ -98,7 +99,7 @@ fn test_lookup_hosts() {
     assert_eq!(lookup.iter().next().unwrap(), Ipv4Addr::new(10, 0, 1, 104));
 }
 
-fn create_ip_like_example() -> Authority {
+fn create_ip_like_example() -> SqliteAuthority {
     let mut authority = create_example();
     authority.upsert(
         Record::new()
@@ -118,7 +119,7 @@ fn create_ip_like_example() -> Authority {
 fn test_lookup_ipv4_like() {
     let authority = create_ip_like_example();
     let mut catalog = Catalog::new();
-    catalog.upsert(authority.origin().clone(), authority);
+    catalog.upsert(authority.origin().clone(), Box::new(authority));
 
     let mut io_loop = Runtime::new().unwrap();
     let (stream, sender) = TestClientStream::new(Arc::new(Mutex::new(catalog)));
@@ -151,7 +152,7 @@ fn test_lookup_ipv4_like() {
 fn test_lookup_ipv4_like_fall_through() {
     let authority = create_ip_like_example();
     let mut catalog = Catalog::new();
-    catalog.upsert(authority.origin().clone(), authority);
+    catalog.upsert(authority.origin().clone(), Box::new(authority));
 
     let mut io_loop = Runtime::new().unwrap();
     let (stream, sender) = TestClientStream::new(Arc::new(Mutex::new(catalog)));
